catalyst_count/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-progressbarupload==0.1.7
djangorestframework==3.14.0
idna==3.4
numpy==1.25.0
oauthlib==3.2.2
panda==0.3.1
pandas==2.0.3
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3